Fixing problem with extensions test, utility tables were added to core
