+++ osquery/tables/specs/osquery_packs.table
